pushhhhhh

+++ app/Http/Controllers/HomeController.php
@@ -976,11 +976,21 @@ class HomeController extends Controller
                 }
 
 
+                if(Auth::user()->hold_wallet < $order->cost){
+                    $hamount = Auth::user()->hold_wallet;
+                    $hamm = number_format($hamount);
+                    $mamm = number_format(Auth::user()->wallet);
+                    return back()->with('message', "Something went wrong | $hamm | $order->cost | $mamm  ");
+                }
+
+
                 if ($corder == 1) {
                     $amount = number_format($order->cost, 2);
 
                     Verification::where('id', $request->id)->delete();
                     User::where('id', Auth::id())->increment('wallet', $order->cost);
+                    User::where('id', Auth::id())->decrement('hold_wallet', $order->cost);
+
 
 
                     $email = User::where('id', $order->user_id)->first()->email ?? null;
